algo/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
gunicorn>=20.1.0
celery>=5.2.0
redis>=4.3.4
sqlalchemy>=1.4.0
pymysql>=1.0.2
alembic>=1.6.0
pydantic-settings>=2.0.0
starlette>=0.27.0 
python-dotenv>=0.19.0
fastapi>=0.95.0,<1.0.0
uvicorn[standard]>=0.22.0
gunicorn>=20.1.0
sqlalchemy[asyncio]>=2.0.0,<3.0.0
asyncio>=0.2.6
# alembic>=1.10.0
redis[hiredis]>=4.3.4
pydantic-settings>=2.0.0
python-dotenv>=0.21.0
starlette>=0.27.0
httpx>=0.24.0
pyprojroot>=0.3.0
cryptography>=41.0.0